+++ src/broadcast.rs
@@ -4,3 +4,9 @@ pub use crate::broadcast_activity_view_controller::{
 pub use crate::broadcast_controller::{
     BroadcastController, BroadcastControllerEvent, BroadcastControllerObserver,
 };
+pub use crate::broadcast_extension::{
+    BroadcastExtensionContext, BroadcastingApplicationInfo,
+    RP_APPLICATION_INFO_BUNDLE_IDENTIFIER_KEY,
+};
+pub use crate::broadcast_handler::BroadcastHandler;
+pub use crate::broadcast_sample_handler::BroadcastSampleHandler;